Optimized initialization speed.
The DEM and land cover datas are now loaded in parallel. Once the
loading is finished, the engine is ready to serve all plugings except
autocomplete.  Hence we start a new thread for initializing the
structures needed by autocomplete, and finish the init-method. This
speeds up the initialization of most plugins considerably.

The engine now has a isSuggestReady() method which is used exclusevily
by the autocomplete plugin to see whether the engine is ready for it
too.

As a result of this optimization many services are ready to serve in
less than 10 seconds even in development machines, while autocomplete
may still take more than a minute to register.